+++ src/main/java/com/ums/service/impl/AdminServiceImpl.java
@@ -53,17 +53,17 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         adminMapper.deleteById(id);
     }
 
     @Override
-    public void deleteBatch(List<Long> ids) {
+    public void deleteBatch(List<Integer> ids) {
         adminMapper.deleteBatch(ids);
     }
 
     @Override
-    public void updateLoginTime(Long id) {
+    public void updateLoginTime(Integer id) {
         adminMapper.updateLoginTime(id);
     }
 }
